+++ CHANGELOG.md
@@ -10,7 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - "Keyple Plugin Stub Library" to version `2.1.0`
 ### Added
 - Calypso example #12 for Performance Measurement (embedded validation).
-- Calypso example #13 for Performance Measurement (distributed reload).
+- Calypso example #13 for Performance Measurement (distributed reloading).
 
 
 ## [2022-06-09]